requirements.txt
fastapi==0.115.0
uvicorn==0.31.0
sqlalchemy==2.0.31
pydantic==2.8.2
openai==1.41.0
python-dotenv==1.0.1
aiofiles==23.2.1
httpx==0.27.0
reportlab==4.1.0
python-multipart==0.0.9
supabase==2.4.0
fpdf==1.7.2
googletrans==4.0.0rc1
